+++ internal/write/queue_test.go
@@ -13,7 +13,9 @@ import (
 func TestQueue(t *testing.T) {
 	que := newQueue(2)
 	assert.True(t, que.isEmpty())
-	b := &Batch{batch: "batch", retryDelay: 3, retryAttempts: 3}
+	assert.Nil(t, que.first())
+	assert.Nil(t, que.pop())
+	b := &Batch{Batch: "batch", RetryDelay: 3, RetryAttempts: 3}
 	que.push(b)
 	assert.False(t, que.isEmpty())
 	b2 := que.pop()